patch: Fix javadoc syntax and outdated refs
This commit fixes some Javadoc problems that prevented compiling the API
docs. Note, however, that there are still loads of outdated references
in this project's Javadoc. This commit is only trying to fix errors that
prevented Javadoc from successfully outputting a compiled API doc, but
the doc comments are still outdated.