accurate-collisions: Fix collision for puddles
This commit disables the accurate block collisions check for puddles of
water. The motion value only has to be accounted for when falling
through a floating layer of water, and also resetting the fall distance
for water that's directly covering the floor incorrectly negates the
damage for a rare client-side glitch that immediately makes you leave
the water you just entered.