nms: Add option to add custom NMS patches
This commit adds a way for us to hook into the very first patching step,
the CraftBukkit nms-patches. This is (luckily) not necessary for most
edits as CraftBukkit already patches most of Minecraft's classes (262 by
default), but for certain modifications, like modifying `BlockBed.java`,
this is the only way to get access to the classes.
As a first major change that requires this patch, this commit includes
a patch for `BlockBed` that moves the (slightly spammy) "You can only
sleep at night" to the action bar.

Potential Issues/Problems

This patch is a major change to Paper's build system. As most Paper
forks won't have the updated `init.sh`, there might be some
complications when blindly copying patches from the PaperSpigot-Server
directory. Thus, patches that depend on the updated `init.sh` should
clearly state this change as a dependency.
Moreover, having a second option to patch CraftBukkit via the
nms-patches directory makes it tempting to apply small modifications
directly using this directory. This, however, is NOT what this directory
was intended and added for. Files in `nms-patches` should be considered
immutable, and you should NEVER edit any files in there. If you want to
edit something in a class that's patched by a file in this directory,
use regular patches in Spigot-Server-Patches instead. You should only
create new files in `nms-patches` if you're trying to patch a class that
doesn't have any CraftBukkit patches yet (or more generally, you should
only use the nms-patches directory if you don't find the vanilla class
you're looking for in `net/minecraft/server` already).

Why this change is still needed

As mentioned above, some classes are not patched by CraftBukkit. As
such, there is no way to change them in PaperSpigot-Server without
copying the entire vanilla class code from `work/`. Needless to say that
this is not an option due to Copyright problems. If used correctly, this
change makes interaction with originally unchanged server classes more
accessible while not imposing any bottlenecks that other solutions like
reflection or packet interception would add.