patch: update entity head yaw on teleport
When /tp is used to teleport an entity, it will also update its head yaw rotation, which is different from the body yaw.

Bukkit's teleport does not mimic this behaviour, so entities are teleported with their old head yaw.

This is not a problem when the entity's AI updates it shortly after. When we have to disable that AI, this is a road blocker.

Some entities (like villagers) may rely on their head yaw to render properly.