nsm-patches: Add outbound packet specialization
This commit adds a simple API that allows plugins to register their
specialized outbound (clientbound) packet classes in EnumProtocol. This
allows the server to assign a packet ID to a non-vanilla packet class.

As mentioned above, this commmit requires the global paper NMS patching
setup, so if you want to include the patch in your fork of Paper, you'll
have to apply the NMS patching setup patch too.