event: Add PlayerTrackEntityEvent
This commit adds an event that gets fired whenever an entity is added to
a player's entity tracker. Cancelling this event prevents the entity
from being added, and consequently, the entity is never sent (nor
updated) for the player. While this is essentially the same method that
Bukkit's hidden player system uses, simply cancelling the event does
not make sure that the entity is not interacted with, and the player can
still see them in some places (e.g. in the player list and the
scoreboard).