build: Update JUnit4 dependency and Guava
This commit updates JUnit4 to prevent possible information exposure with
the `TemporaryFolder` test rule. The project currently uses this rule,
but due to the nature of a public free software project, the data
exposed by this vulnerability is not of great value to a potential local
attacker. As the dependency was declared with scope `test`, running
servers weren't affected by the vulnerability.

CVE: CVE-2020-15250
GHSA: GHSA-269g-pwp5-87pp

Additionally, this commit updates Guava to prevent two possible
vulnerabilities:

Deserializing attacker-provided data could lead to a denial of service
because `AtomicDoubleArray` when serialized with Java serialization and
`CompoundOrdering` when serialized with GWT serialization performed no
appropriate checks whether the user-provided data has a reasonable size.
KigPaper doesn't seem affected by this vulnerability, but given that
Guava is a runtime dependency, it should still be updated.

CVE: CVE-2018-10237

Temporary directories created with
`com.google.common.io.Files#createTempDir()` would be readable by an
attacker with local access to the system. While updating the dependency
does *not* fix this security vulnerability, the method has been
deprecated in Guava 30. It is advised to "use
`java.nio.file.Files#createTempDirectory`, transforming it to a `File`
using `java.nio.file.Path#toFile()` if needed". KigPaper doesn't seem
affected by this vulnerability as the method in question is never
invoked. Nevertheless, this dependency should be updated to hint the
projects reusing the server's Guava library to switch to the safe
alternative mentioned in the deprecation notice (and quoted in this
commit).

CVE: CVE-2020-8908