nms: Upgrade Guava in untouched NMS classes
This commit adds various NMS classes that referenced the now updated
version of Guava to the nms-patches system with an initial patch that
updates the classes to the new Guava version.

Adding the classes to `nms-patches` is necessary as they weren't changed
by CraftBukkit. Thus, they're only available if we use the custom patch
override.

Without the patches in this commit, the server would throw exceptions
when generating terrain or performing block state changes.